Update .gitignore, add new requirements, enable MySQL integration
In this commit, we've updated the .gitignore file and added new libraries to fulfill the requirement.txt file. By adding these new libraries we're enabling the bot to integrate with a MySQL database for better score tracking and reporting. This also improves error handling in discord embed messages for user feedback. Features like a 'leaderboard' command have been added to this update