Add Version 4.0 details to changelog embed
In the latest version of the Discord bot, a new field was added to the changelog embed to list all the changes in Version 4.0. These changes include a new scaling system, a refactored infrastructure, changes to the leaderboard display, balancing item prices, inclusion of a "Russian Roulette" item, and a fix for an overflow issue in the webapp.